+++ tests/test_simulatedorder.py
@@ -994,7 +994,7 @@ class SimulatedOrderTest(unittest.TestCase):
         "flumine.simulation.simulatedorder.SimulatedOrder._calculate_process_traded"
     )
     def test__process_traded_back(self, mock__calculate_process_traded):
-        mock__calculate_process_traded.return_value = 5.
+        mock__calculate_process_traded.return_value = 5.0
         traded = {12: 120}
         self.simulated._process_traded(1234579, traded)
         mock__calculate_process_traded.assert_called_with(1234579, 120)
@@ -1006,7 +1006,7 @@ class SimulatedOrderTest(unittest.TestCase):
     def test__process_traded_over_fill(self, mock__calculate_process_traded):
         for side in ("BACK", "LAY"):
             self.simulated.order.side = side
-            mock__calculate_process_traded.return_value = 130.
+            mock__calculate_process_traded.return_value = 130.0
             traded = {12: 120}
             self.simulated._process_traded(1234579, traded)
             mock__calculate_process_traded.assert_called_with(1234579, 120)
@@ -1023,7 +1023,7 @@ class SimulatedOrderTest(unittest.TestCase):
         "flumine.simulation.simulatedorder.SimulatedOrder._calculate_process_traded"
     )
     def test__process_traded_lay(self, mock__calculate_process_traded):
-        mock__calculate_process_traded.return_value = 10.
+        mock__calculate_process_traded.return_value = 10.0
         self.simulated.order.side = "LAY"
         traded = {12: 120}
         self.simulated._process_traded(1234581, traded)